requirements.txt
bibtexparser
pyperclip
requests
textrank4zh
jieba
numpy
scipy
# nltk
summa
pytoml

#format
autopep8

#translator
googletrans
tencentcloud_sdk_python